added Django-autocomplete-light to add user to project form
required the installation of Django-autocomplete-light (via pip)

attempted installing jquery directly, but during troubleshooting
reverted to the online script already in place. Also installed
select2 (select2.org), but removed from the html, as it doesn't
seem to be required (functions are likely present in the
dal_select2 module).

requires modification of the settings.py file to include dal and
dal_select2 (add before the adminconfig app) in INSTALLED_APPS.